Composer can handle different indentations
Historically, Composer would add new lines (via `composer require...`) with a hard-coded 4-space indentation.

This is no longer the case, and it will match whatever the existing indentation is.

As such, we no longer need to specify 4 spaces, so it can default to being 2 spaces as per the de facto standard for JSON files.