CS: Rector with Deadcode:true
RemoveUselessReturnTagRector
RemoveUselessVarTagRector

With the "useless" var tags removed, the `Squiz.Commenting.VariableComment` sniff used for the WordPress-Docs standard complains that the expected var tag is not found. In this case, I'm happy to make an opinionated call and just ignore this. The PHPDocs are unlikely to ever by extracted from a plugin built with this boilerplate, and even if it is, the parser should be updated to read from the type declarations instead.